+++ js/utils/debounce.js
@@ -1,21 +1,16 @@
 // Функция взята из интернета и доработана
 // Источник - https://www.freecodecamp.org/news/javascript-debounce-example
-
+import {cleanMiniatures} from './clean-miniatures.js';
 function debounce (callback, timeoutDelay) {
-  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
-  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
   let timeoutId;
 
   return (...rest) => {
-    // Перед каждым новым вызовом удаляем предыдущий таймаут,
-    // чтобы они не накапливались
     clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => {
+      cleanMiniatures();
+      callback.apply(this, rest);
+    }, timeoutDelay);
 
-    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
-    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-
-    // Таким образом цикл "поставить таймаут - удалить таймаут" будет выполняться,
-    // пока действие совершается чаще, чем переданная задержка timeoutDelay
   };
 }
